Add gulp_installVSCodeExtension
