+++ dev/src/index.ts
@@ -9,4 +9,5 @@ export * from './TestOutputChannel';
 export * from './TestUserInput';
 export { getDefaultWebpackConfig } from './webpack/getDefaultWebpackConfig';
 export * from './gulp/gulp_installAzureAccount';
+export * from './gulp/gulp_installVSCodeExtension';
 export * from './gulp/gulp_webpack';
